Grammatical changes for cert approval
Re-worded some of the sentences because 'email approval' in
some cases might be too detailed, but 'email approval process' is a
better description of the overall task here.

Also corrected preposition usage and softened the Whois privacy sentence
with suggestions of 'temporarily' disabling instead of people
potentially misinterpreting it as permanent.