Use consistent instructions
Normalize the way we tell the user to reach account-specific information.

We also used the term account when referencing the login, but you don't log into the account. You login with your user credentials, then select an account.

Also avoid using the dashboard as shortcut, that creates confusion compared to other articles.

I am aware we can use the search for the domain. But if this is the approach, the message must be consistent in all articles.