Duplicated search bar (#377)
Closes GH-273

I've deleted the extra search bar, as the one on the header is common for every page.
I've change the text a bit because it was about searching on the extra search bar.

I've highlighted the header search bar adding the inner shadow we use on the app, and changing the button from default to primary.

I've taken this opportunity to change back the h1 color to dark grey instead of blue, because we are using blue for links. This also helps make more visible the blue from the button.