Call out 5-day waiting period to transfer domain
I encountered a light "gotcha" situation, in that it may not be
clear to customers that transfers into DNSimple can take up to 5 days.
This change identifies that as a caveat in the docs and links to an old
blog post that gives a bit more context as to why that's the case.

- https://support.dnsimple.com/articles/transferring-domain/
- https://blog.dnsimple.com/2012/01/things-to-know-about-transferring-a-domain/
- https://blog.dnsimple.com/2011/02/preparing-transfer-from-godaddy/